Implement data and domain madule and done code formatting
