--- app/src/main/java/com/example/comicapp/data/ComicDataSource.kt
@@ -1,8 +1,0 @@
-package com.example.comicapp.data
-
-import com.example.comicapp.home.ComicModel
-import com.example.comicapp.util.UiState
-
-interface ComicDataSource {
-    suspend fun getComicList(): UiState<List<ComicModel>>
-}
