--- app/src/main/java/com/example/comicapp/data/ComicRepository.kt
@@ -1,9 +1,0 @@
-package com.example.comicapp.data
-
-import com.example.comicapp.home.ComicModel
-import com.example.comicapp.util.UiState
-
-interface ComicRepository {
-
-    suspend fun getComicList(): UiState<List<ComicModel>>
-}
