--- app/src/main/java/com/example/comicapp/data/Mapper.kt
@@ -1,6 +1,0 @@
-package com.example.comicapp.data
-
-abstract class Mapper<in E, T> {
-
-    abstract fun mapTo(from: E): T
-}
