--- app/src/main/java/com/example/comicapp/home/ComicModel.kt
@@ -1,6 +1,0 @@
-package com.example.comicapp.home
-
-data class ComicModel(
-    val imageUrl: String?,
-    val comicName: String?,
-)
