+++ common/src/main/java/com/example/common/data/dispatcher/CoroutineDispatcherProvider.kt
@@ -1,0 +1,20 @@
+package com.example.common.data.dispatcher
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers.Default
+import kotlinx.coroutines.Dispatchers.IO
+import kotlinx.coroutines.Dispatchers.Main
+import javax.inject.Inject
+
+data class CoroutineDispatcherProvider(
+    val main: CoroutineDispatcher,
+    val io: CoroutineDispatcher,
+    val default: CoroutineDispatcher,
+) {
+    @Inject
+    constructor() : this(
+        Main,
+        IO,
+        Default
+    )
+}
