+++ data/src/main/java/com/example/data/remote/ComicApi.kt
@@ -1,0 +1,14 @@
+package com.example.data.remote
+
+import com.example.data.dto.ComicDto
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ComicApi {
+
+    @GET("comics")
+    suspend fun getComicList(
+        @Query("offset") offset: Int = 0,
+        @Query("limit") limit: Int = 20,
+    ): ComicDto?
+}
