+++ domain/src/main/java/com/example/domain/ComicDataSource.kt
@@ -1,0 +1,7 @@
+package com.example.domain
+
+import com.example.domain.model.ComicModel
+
+interface ComicDataSource {
+    suspend fun getComicList(): Result<List<ComicModel>>
+}
