+++ domain/src/main/java/com/example/domain/ComicRepository.kt
@@ -1,0 +1,8 @@
+package com.example.domain
+
+import com.example.domain.model.ComicModel
+
+interface ComicRepository {
+
+    suspend fun getComicList(): Result<List<ComicModel>>
+}
