+++ domain/src/main/java/com/example/domain/Mapper.kt
@@ -1,0 +1,6 @@
+package com.example.domain
+
+abstract class Mapper<in E, T> {
+
+    abstract fun mapTo(from: E): T
+}
