+++ domain/src/main/java/com/example/domain/model/ComicModel.kt
@@ -1,0 +1,6 @@
+package com.example.domain.model
+
+data class ComicModel(
+    val imageUrl: String?,
+    val comicName: String?,
+)
