+++ domain/src/main/java/com/example/domain/usecase/BaseUseCase.kt
@@ -1,0 +1,4 @@
+package com.example.domain.usecase
+
+interface BaseUseCase {
+}
