+++ settings.gradle
@@ -17,3 +17,5 @@ include ':app'
 include ':common'
 include ':commonui'
 include ':home'
+include ':data'
+include ':domain'
